pkg/types: Push platform-specific types (AWS, etc.) into subdirs
This decouples our platforms a bit and makes it easier to distinguish
between platform-specific and platform-agnostic code.  It also gives
us much more compact struct names, since now we don't need to
distinguish between many flavors of machine pool, etc. in a single
package.

I've also updated pkg/types/doc.go; pkg/types includes more than
user-specified configuration since 78c31183 (pkg: add ClusterMetadata
asset,type that can be used for destroy, 2018-09-25, #324).

I've also added OWNERS files for some OpenStack-specific directories
that were missing them before.

There's still more work to go in this direction (e.g. pushing default
logic into subdirs), but this seems like a reasonable chunk.
