+++ pkg/types/aws/doc.go
@@ -1,0 +1,3 @@
+// Package aws contains AWS-specific structures for installer
+// configuration and management.
+package aws
