+++ pkg/types/aws/machinepool.go
@@ -1,0 +1,62 @@
+package aws
+
+// MachinePool stores the configuration for a machine pool installed
+// on AWS.
+type MachinePool struct {
+	// Zones is list of availability zones that can be used.
+	Zones []string `json:"zones,omitempty"`
+
+	// AMIID defines the AMI that should be used.
+	AMIID string `json:"amiID,omitempty"`
+
+	// InstanceType defines the ec2 instance type.
+	// eg. m4-large
+	InstanceType string `json:"type"`
+
+	// IAMRoleName defines the IAM role associated
+	// with the ec2 instance.
+	IAMRoleName string `json:"iamRoleName"`
+
+	// EC2RootVolume defines the storage for ec2 instance.
+	EC2RootVolume `json:"rootVolume"`
+}
+
+// Set sets the values from `required` to `a`.
+func (a *MachinePool) Set(required *MachinePool) {
+	if required == nil || a == nil {
+		return
+	}
+
+	if len(required.Zones) > 0 {
+		a.Zones = required.Zones
+	}
+	if required.AMIID != "" {
+		a.AMIID = required.AMIID
+	}
+	if required.InstanceType != "" {
+		a.InstanceType = required.InstanceType
+	}
+	if required.IAMRoleName != "" {
+		a.IAMRoleName = required.IAMRoleName
+	}
+
+	if required.EC2RootVolume.IOPS != 0 {
+		a.EC2RootVolume.IOPS = required.EC2RootVolume.IOPS
+	}
+	if required.EC2RootVolume.Size != 0 {
+		a.EC2RootVolume.Size = required.EC2RootVolume.Size
+	}
+	if required.EC2RootVolume.Type != "" {
+		a.EC2RootVolume.Type = required.EC2RootVolume.Type
+	}
+}
+
+// EC2RootVolume defines the storage for an ec2 instance.
+type EC2RootVolume struct {
+	// IOPS defines the iops for the instance.
+	IOPS int `json:"iops"`
+	// Size defines the size of the instance.
+	Size int `json:"size"`
+	// Type defines the type of the instance.
+	Type string `json:"type"`
+}
