+++ pkg/types/doc.go
@@ -1,2 +1,3 @@
-// Package types defines structures for user-supplied installer configuration.
+// Package types defines structures for installer configuration and
+// management.
 package types
