+++ pkg/types/libvirt/doc.go
@@ -1,0 +1,3 @@
+// Package libvirt contains libvirt-specific structures for
+// installer configuration and management.
+package libvirt
