+++ pkg/types/libvirt/metadata.go
@@ -1,0 +1,6 @@
+package libvirt
+
+// Metadata contains libvirt metadata (e.g. for uninstalling the cluster).
+type Metadata struct {
+	URI string `json:"uri"`
+}
