+++ pkg/types/libvirt/network.go
@@ -1,0 +1,11 @@
+package libvirt
+
+// Network is the configuration of the libvirt network.
+type Network struct {
+	// Name is the name of the nework.
+	Name string `json:"name"`
+	// IfName is the name of the network interface.
+	IfName string `json:"if"`
+	// IPRange is the range of IPs to use.
+	IPRange string `json:"ipRange"`
+}
