+++ pkg/types/openstack/doc.go
@@ -1,0 +1,3 @@
+// Package openstack contains OpenStack-specific structures for
+// installer configuration and management.
+package openstack
