+++ pkg/types/openstack/metadata.go
@@ -1,0 +1,8 @@
+package openstack
+
+// Metadata contains OpenStack metadata (e.g. for uninstalling the cluster).
+type Metadata struct {
+	Region string `json:"region"`
+	// Most OpenStack resources are tagged with these tags as identifier.
+	Identifier map[string]string `json:"identifier"`
+}
